corection

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -7,9 +7,9 @@
  */
 void print_number(int n)
 {
-	long m; /* power of 10 */
+	int m; /* power of 10 */
 	int c; /* boolean check */
-	long num; /* convert int to long */
+	int num; /* int  */
 
 	num = n;
 	/* negatives */
